Allow for repeated and/or unsorted entries (allow periodic dimensions)
This enables support for periodic dimensions.

Also remove option for ranges to be UnitRange (internally, inputs are
still allowed to be `UnitRange{Int64}`, but these are converted to
`Vector{Int64}`). Might be nice to have the option of UnitRange for
cases where it is allowed (probably a bit more optimal), but makes code
more complicated, and in the use cases we are interested in we probably
have to use Vector{Int64} anyway.